uno/requirements.txt
pinecone
python-dotenv